chore: rename to MediaItem
