+++ libs/ui/src/components/MediaItem/MediaItem.vue
