+++ libs/ui/src/components/MediaItem/type/AudioMedia.vue
