+++ libs/ui/src/components/MediaItem/type/ImageMedia.vue
