+++ libs/ui/src/components/MediaItem/type/JsonMedia.vue
