+++ libs/ui/src/components/MediaItem/type/ModelMedia.vue
