+++ libs/ui/src/components/MediaItem/type/ObjectMedia.vue
