+++ libs/ui/src/index.ts
@@ -1,4 +1,4 @@
 export { default as CardArticle } from './components/CardArticle/CardArticle.vue'
-export { default as MediaObject } from './components/MediaObject/MediaObject.vue'
+export { default as MediaItem } from './components/MediaItem/MediaItem.vue'
 
 export { default as SampleButton } from './components/TheButton/TheButton.vue'
